reorg log structure; cleanup logging
