+++ requirements.txt
@@ -17,6 +17,7 @@ six==1.11.0
 urllib3==1.24.2
 Werkzeug==0.14.1
 dateparser==0.7.1
-requests_oauthlib
+requests_oauthlib==0.8.0
+oauthlib==3.0.0
 PyPDF2==1.26.0
 tika==1.19
